% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deciduous_polygon}
\alias{deciduous_polygon}
\title{128 NDVI pixels from a MOD13Q1 time series}
\format{
An object of class \code{matrix}.
}
\usage{
data(deciduous_polygon)
}
\description{
Small spatial subset of a MOD13Q1 time series from 2000 to 2021.
The MOD13Q1 provides measurements of the Normalized Difference Vegetation 
Index (NDVI), a variable that is suitable to conduct mid-term vegetation studies remotely.
The pixels provided by this dataset were recorded from a deciduous forest zone.
}
\details{
The dataset is distributed through an \code{RData} file containing a 
\code{matrix} object with 128 rows and 506 columns.
}
\keyword{datasets}
