% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDist_phenoParam.R
\name{getDist_phenoParam}
\alias{getDist_phenoParam}
\title{Utility function}
\usage{
getDist_phenoParam(
  LIST,
  phenoParam = c("GU", "SoS", "Mat", "Sen", "EoS", "Dor")
)
}
\arguments{
\item{LIST}{list, containing 6 estimated phenological parameters}

\item{phenoParam}{character. What phenological parameter should be extracted?}
}
\value{
A numeric vector
}
\description{
Extracts an estimated phenological parameter from a \code{list}. Useful when 
\code{phenopar_polygon} was applied to estimate phenological dates over a polygon.
}
\seealso{
\code{\link[sephora]{getSpiralPlot}}, \code{\link[sephora]{phenopar_polygon}}
}
