% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpiralPlot.R
\name{getSpiralPlot}
\alias{getSpiralPlot}
\title{Spiral plot of phenological parameters}
\usage{
getSpiralPlot(LIST, MAT = NULL, height = 0.2, LABELS, ...)
}
\arguments{
\item{LIST}{list, containing 6 estimated phenological parameters.}

\item{MAT}{matrix, containing 6 estimated phenological parameters. Default, \code{NULL}.}

\item{height}{numeric, height parameter of \code{\link[spiralize]{spiral_track}} (used internally)}

\item{LABELS}{character, labels parameter of \code{\link[spiralize]{spiral_axis}} (used internally)}

\item{...}{additional parameters to \code{\link[spiralize]{spiral_initialize}}}
}
\value{
No value is returned
}
\description{
This utility function yields a spiral plot based on
phenological dates estimated from a polygon.
}
\seealso{
\code{\link[sephora]{getSpiralPlot}}, \code{\link[sephora]{phenopar_polygon}}, \code{\link[spiralize]{spiral_track}}, 
\code{\link[spiralize]{spiral_axis}}, \code{\link[spiralize]{spiral_initialize}}
}
