% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sephora-class.R
\docType{class}
\name{sephora-class}
\alias{sephora-class}
\title{class sephora}
\description{
Definition of the \code{sephora} class
}
\section{Slots}{

\describe{
\item{\code{x}}{Original time series (as a numeric vector)}

\item{\code{startYear}}{Beginning of time series}

\item{\code{endYear}}{End of time series}

\item{\code{freq}}{Number of observations per season}

\item{\code{sigma}}{Variability estimate}

\item{\code{m_aug_smooth}}{Samples of smoothed version of \code{x}, in matricial form}

\item{\code{clustering}}{An object of class \code{\link[dtwclust]{HierarchicalTSClusters}}}

\item{\code{fpca}}{Numeric, FPCA-based regression fit}

\item{\code{fpca_harmfit_params}}{a list, harmonic fit}

\item{\code{fpca_fun_0der}}{Function fpca fit}

\item{\code{fpca_fun_1der}}{Function fpca fit first derivative}

\item{\code{fpca_fun_2der}}{Function fpca fit second derivative}

\item{\code{fpca_fun_3der}}{Function fpca fit third derivative}

\item{\code{fpca_fun_4der}}{Function fpca fit fourth derivative}

\item{\code{phenoparams}}{Phenological dates estimate}

\item{\code{status}}{Character, was phenopar estimation successful?}
}}

\seealso{
\code{\link[sephora]{sephora-methods}}
}
