% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecToMatrix.R
\name{vecToMatrix}
\alias{vecToMatrix}
\title{Mapping numeric vector to a matrix}
\usage{
vecToMatrix(x, lenPeriod = 23)
}
\arguments{
\item{x}{a numeric vector whose length must be a multiple of \code{lenPeriod}}

\item{lenPeriod}{a numeric, number of observations per period}
}
\value{
A matrix with \code{nrow} equal to \code{length(x)/lenPeriod} and
\code{ncol} equal to \code{lenPeriod}.
}
\description{
Maps a vector (pixel of a satellite time series) to a matrix.
}
\seealso{
\code{\link[sephora]{fill_initialgap_MOD13Q1}}, \code{\link[sephora]{phenopar}}, 
\code{\link[sephora]{vecFromData}}.
}
